+++ conf.lua
@@ -1,4 +1,4 @@
 function love.conf(t)
-	t.window.title = "idops v0.0.4 demo"
+	t.window.title = "idops v0.0.5 demo"
 	t.window.resizable = true
 end